+++ internal/mock/conn.go
@@ -14,7 +14,7 @@ type Conn struct {
 	PingErr       error
 	AddrField     string
 	LocationField string
-	NodeIDField   uint32
+	NodeIDField   int64
 	State         conn.State
 	LocalDCField  bool
 }
@@ -80,14 +80,14 @@ func (c *Conn) Unban(ctx context.Context) conn.State {
 type Endpoint struct {
 	AddrField     string
 	LocationField string
-	NodeIDField   uint32
+	NodeIDField   int64
 	LocalDCField  bool
 }
 
 func (e *Endpoint) Choose(bool) {
 }
 
-func (e *Endpoint) NodeID() uint32 {
+func (e *Endpoint) NodeID() int64 {
 	return e.NodeIDField
 }