+++ internal/table/transaction.go
@@ -25,15 +25,20 @@ var (
 )
 
 type txState struct {
-	rawVal atomic.Uint32
+	atomic.Pointer[txStateEnum]
 }
 
-func (s *txState) Load() txStateEnum {
-	return txStateEnum(s.rawVal.Load())
+func (s *txState) Get() txStateEnum {
+	ptr := s.Pointer.Load()
+	if ptr == nil {
+		return txStateInitialized
+	}
+
+	return *ptr
 }
 
-func (s *txState) Store(val txStateEnum) {
-	s.rawVal.Store(uint32(val))
+func (s *txState) Set(state txStateEnum) {
+	s.Pointer.Store(&state)
 }
 
 type txStateEnum uint32
@@ -52,6 +57,10 @@ type transaction struct {
 }
 
 func (tx *transaction) ID() string {
+	if tx == nil {
+		return ""
+	}
+
 	return tx.id
 }
 
@@ -70,7 +79,7 @@ func (tx *transaction) Execute(
 		onDone(r, err)
 	}()
 
-	switch tx.state.Load() {
+	switch tx.state.Get() {
 	case txStateCommitted:
 		return nil, xerrors.WithStackTrace(errTxAlreadyCommitted)
 	case txStateRollbacked:
@@ -82,7 +91,7 @@ func (tx *transaction) Execute(
 		}
 
 		if tx.control.Desc().GetCommitTx() {
-			tx.state.Store(txStateCommitted)
+			tx.state.Set(txStateCommitted)
 		}
 
 		return r, nil
@@ -107,7 +116,7 @@ func (tx *transaction) ExecuteStatement(
 		onDone(r, err)
 	}()
 
-	switch tx.state.Load() {
+	switch tx.state.Get() {
 	case txStateCommitted:
 		return nil, xerrors.WithStackTrace(errTxAlreadyCommitted)
 	case txStateRollbacked:
@@ -119,7 +128,7 @@ func (tx *transaction) ExecuteStatement(
 		}
 
 		if tx.control.Desc().GetCommitTx() {
-			tx.state.Store(txStateCommitted)
+			tx.state.Set(txStateCommitted)
 		}
 
 		return r, nil
@@ -140,7 +149,7 @@ func (tx *transaction) CommitTx(
 		onDone(err)
 	}()
 
-	switch tx.state.Load() {
+	switch tx.state.Get() {
 	case txStateCommitted:
 		return nil, xerrors.WithStackTrace(errTxAlreadyCommitted)
 	case txStateRollbacked:
@@ -167,7 +176,7 @@ func (tx *transaction) CommitTx(
 			}
 		}
 
-		response, err = tx.s.tableService.CommitTransaction(ctx, request)
+		response, err = tx.s.client.CommitTransaction(ctx, request)
 		if err != nil {
 			return nil, xerrors.WithStackTrace(err)
 		}
@@ -177,7 +186,7 @@ func (tx *transaction) CommitTx(
 			return nil, xerrors.WithStackTrace(err)
 		}
 
-		tx.state.Store(txStateCommitted)
+		tx.state.Set(txStateCommitted)
 
 		return scanner.NewUnary(
 			nil,
@@ -198,13 +207,13 @@ func (tx *transaction) Rollback(ctx context.Context) (err error) {
 		onDone(err)
 	}()
 
-	switch tx.state.Load() {
+	switch tx.state.Get() {
 	case txStateCommitted:
 		return nil // nop for committed tx
 	case txStateRollbacked:
 		return xerrors.WithStackTrace(errTxRollbackedEarly)
 	default:
-		_, err = tx.s.tableService.RollbackTransaction(ctx,
+		_, err = tx.s.client.RollbackTransaction(ctx,
 			&Ydb_Table.RollbackTransactionRequest{
 				SessionId: tx.s.id,
 				TxId:      tx.id,
@@ -220,7 +229,7 @@ func (tx *transaction) Rollback(ctx context.Context) (err error) {
 			return xerrors.WithStackTrace(err)
 		}
 
-		tx.state.Store(txStateRollbacked)
+		tx.state.Set(txStateRollbacked)
 
 		return nil
 	}