+++ metrics/node_id.go
@@ -4,6 +4,6 @@ import (
 	"strconv"
 )
 
-func idToString(id uint32) string {
-	return strconv.FormatUint(uint64(id), 10)
+func idToString(id int64) string {
+	return strconv.FormatInt(id, 10)
 }